Create rdconfig yaml along with SNAT CR
If clusterSNATPolicyIP field is set, we add rdconfig file called cluster-network-{{ rdconfigFileNumber }}-2-CustomResource-rdconfig.yaml like this:
apiVersion: aci.snat/v1
kind: RdConfig
metadata:
  name: routingdomain-config
  namespace: aci-containers-system
spec:
  usersubnets:
  - {{ neutronCIDR }}
  - 224.0.0.0/4

The rdconfigFileNumber is decided based on the requirement that we need to create it after the rdconfig CRD is created and a little before the aci-containers-host daemonset is created.
So we find out the filenumber for the DaemonSet, subtract 4 and use that number for rdconfig CR.

The code returns an error if the acc-provision tar doesn't have a file with "DaemonSet-aci-containers-host" in its name.
Caveat: We assume there's a file inside tar with a name of this format: cluster-network-{{ filenumber }}-DaemonSet-aci-containers-host.yaml